refactor(shared): modularize anthropic provider metadata
